+++ src/main/java/com/danrus/rpf/compat/RpfCompatInitializer.java
@@ -1,0 +1,10 @@
+//
+// Source code recreated from a .class file by IntelliJ IDEA
+// (powered by Fernflower decompiler)
+//
+
+package com.danrus.rpf.compat;
+
+public interface RpfCompatInitializer {
+    void init();
+}